newcode

+++ lib/app/models/dashboard_model.dart
@@ -4,7 +4,7 @@ import 'package:hallo_doctor_doctor_app/app/models/timeslot_model.dart';
 class DashboardModel {
   DashboardModel({this.balance, this.listReview, this.listAppointment});
 
-  int? balance;
+  double? balance;
   List<TimeSlot>? listAppointment;
   List<ReviewModel>? listReview;
 }
